+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## WIP v2.0.0
 
+- separate decorators XmlAttribute, XmlChardata, XmlComments, XmlChildElem
+- add basic support for XML comments
+- add support for bigint as primitive type
+- rename XmlEntity to XmlElem, XmlProperty to XmlChildElem
+- concatenate text nodes when parsing chardata (before this only first text node was being parsed)
+- docs: added more JSDocs
+- internal: minor improvements and refactoring
+
 - support circular dependencies by specifying `type` and `union` only via functions
 - rename `AnyClass` to `XmlClass` and add comments about constructor arguments
 - internal: fix `serializeUnionForLog` error messages